+++ module-service/src/main/java/com/welcommu/moduleservice/company/CompanyServiceImpl.java
@@ -5,10 +5,10 @@ import com.welcommu.modulecommon.exception.CustomException;
 import com.welcommu.moduledomain.company.Company;
 import com.welcommu.moduledomain.user.User;
 import com.welcommu.moduleinfra.user.UserRepository;
+import com.welcommu.moduleservice.company.audit.CompanyAuditService;
 import com.welcommu.moduleservice.company.dto.CompanyModifyRequest;
 import com.welcommu.moduleservice.company.dto.CompanyRequest;
 import com.welcommu.moduleinfra.company.CompanyRepository;
-import com.welcommu.moduleservice.company.audit.CompanyAuditService;
 import com.welcommu.moduleservice.company.dto.CompanyResponse;
 import com.welcommu.moduleservice.company.dto.CompanySnapshot;
 import com.welcommu.moduleservice.user.dto.UserResponse;
